requirements.txt
# Web Framework
Flask==2.3.3
Werkzeug==2.3.7

# PDF Processing
pdfplumber==0.10.2
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.2.0

# AI Providers
openai==1.12.0
anthropic==0.18.1
google-generativeai==0.4.1
aiohttp==3.9.1

# Data Export
openpyxl==3.1.2

# Security
cryptography==42.0.2

# Build Tools
pyinstaller==6.3.0

# Async Support
asyncio-compat==0.1.1
